F1 key does not work on modern Windows 11
Relates to #44

This issue is caused by:
https://github.com/YaSuenag/SimpleCom/commit/939f11d03852a2c2f5818afa18e7dfbf933592c6

This change worked at the timing of the commit (Jan 6 2024) at least,
but it does not work at June 22 2024 on Windows 11 23H2 (22631.3737).
So I guess this bug is closely related to Windows updates between
Jan 2024 and June 2024, but I could not figure out the KB which relates
to this.